Understanding Web Automation using Selenium
This Python script demonstrates web automation using the Selenium framework for accessing and interacting with web pages. It navigates to a specified URL (https://edube.org/login), locates a registration link element on the page using XPath, and performs a series of actions, including clicking the registration link. The script uses the Chrome web browser for automation and employs the WebDriver Manager to handle browser driver installations. Explore the capabilities of Selenium for automated web interactions and streamline user registration processes with this efficient and reliable script.